requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
py-zabbix==1.1.7
pycodestyle==2.8.0
python-dateutil==2.8.2
requests==2.26.0
semantic-version==2.8.5
six==1.16.0
toml==0.10.2
urllib3==1.26.7